build: refactor base config
