+++ packages/snap-event-manager/package.json
@@ -16,8 +16,8 @@
 		"dev": "tsc --watch",
 		"format": "prettier --write 'src/**/*.{js,jsx,ts,tsx}'",
 		"lint": "eslint 'src/**/*.{js,jsx,ts,tsx}'",
-		"test": "jest --passWithNoTests --projects=''",
-		"test:watch": "jest --watch --projects=''"
+		"test": "jest --passWithNoTests",
+		"test:watch": "jest --watch"
 	},
 	"sideEffects": false,
 	"files": [
