+++ packages/snap-preact-components/jest.config.js
@@ -1,13 +1,12 @@
-const rootConfig = require('../../jest.config.json');
+const rootConfig = require('../../jest.base.config.json');
 module.exports = {
 	...rootConfig,
+	displayName: 'snap-preact-components',
 	moduleNameMapper: {
 		'^react$': 'preact/compat',
 		'^react-dom/test-utils$': 'preact/test-utils',
 		'^react-dom$': 'preact/compat',
 		'\\.(css|less|sass|scss)$': '<rootDir>/__mocks__/styleMock.js',
 	},
-	roots: ['./src/components'],
 	setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
-	displayName: 'snap-preact-components',
 };
