+++ packages/snap-shared/package.json
@@ -11,8 +11,8 @@
 		"dev": "tsc --watch",
 		"format": "prettier --write 'src/**/*.{js,jsx,ts,tsx}'",
 		"lint": "eslint 'src/**/*.{js,jsx,ts,tsx}'",
-		"test": "jest --passWithNoTests --projects=''",
-		"test:watch": "jest --watch --projects=''"
+		"test": "jest --passWithNoTests",
+		"test:watch": "jest --watch"
 	},
 	"devDependencies": {
 		"@searchspring/snap-client": "^0.21.0"
